+++ src/config.js
@@ -2,34 +2,41 @@
 import createTmplRule from './utils/createTmplRule';
 
 export default function (configs) {
-  const delimiters = configs.delimiters,
-    includeParser = configs.includeParser,
-    createElement = configs.createElement,
-    outputH = configs.outputH,
-    textMode = configs.textMode,
-    noWsMode = configs.noWsMode;
+  const {
+    delimiters,
+    includeParser,
+    createElement,
+    outputH,
+    textMode,
+    noWsMode,
+    fixTagName
+  } = configs;
 
-  if(delimiters) {
+  if (delimiters) {
     createTmplRule(delimiters, true);
   }
 
-  if(includeParser) {
+  if (includeParser) {
     nj.includeParser = includeParser;
   }
 
-  if(createElement) {
+  if (createElement) {
     nj.createElement = createElement;
   }
 
-  if(outputH != null) {
+  if (outputH != null) {
     nj.outputH = outputH;
   }
 
-  if(textMode != null) {
+  if (textMode != null) {
     nj.textMode = textMode;
   }
 
-  if(noWsMode != null) {
+  if (noWsMode != null) {
     nj.noWsMode = noWsMode;
+  }
+
+  if (fixTagName != null) {
+    nj.fixTagName = fixTagName;
   }
 };